order/src/main/java/com/readingisgood/order/service/authentication/AuthenticationService.java
package com.readingisgood.order.service.authentication;

import com.readingisgood.order.request.AuthenticateRequest;
import com.readingisgood.order.response.AuthenticateResponse;
import com.readingisgood.order.security.JwtUtil;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.stereotype.Service;

/** @author hakan.ozerden */
@Service
@Slf4j
@RequiredArgsConstructor
public class AuthenticationService {

    private final AuthenticationManager authenticationManager;

    private final JwtUtil jwtUtil;

    public AuthenticateResponse authenticateUser(AuthenticateRequest request) {

        Authentication authentication =
                authenticationManager.authenticate(
                        new UsernamePasswordAuthenticationToken(
                                request.getUsername(), request.getPassword()));

        SecurityContextHolder.getContext().setAuthentication(authentication);
        String jwt = jwtUtil.generateJwt(authentication);

        return AuthenticateResponse.builder().token(jwt).build();
    }
}
